# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rolls API",
    "version": "1.0.0",
    "description": "OpenAPI spec generated from Next.js route handlers."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local dev (API base)"
    },
    {
      "url": "https://rolls.yannick.computer/api",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/films": {
      "get": {
        "operationId": "get-films",
        "summary": "List films",
        "description": "",
        "tags": [
          "Films"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Film"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-films",
        "summary": "Create (or upsert) a film",
        "description": "Slug is generated from brand+name. Upserts on (user_id, slug).",
        "tags": [
          "Films"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFilmBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Film"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/rolls": {
      "get": {
        "operationId": "get-rolls",
        "summary": "List rolls",
        "description": "Returns all rolls for the authenticated user, ordered by roll_number desc.",
        "tags": [
          "Rolls"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Roll"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-rolls",
        "summary": "Create a roll",
        "description": "Creates a new roll for the authenticated user. camera_id and film_id are slugs.",
        "tags": [
          "Rolls"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRollBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roll"
                }
              }
            }
          }
        }
      }
    },
    "/rolls/{id}/contact-sheet": {
      "get": {
        "operationId": "get-rolls-{id}-contact-sheet",
        "summary": "Get a roll contact sheet",
        "description": "Returns the roll contact sheet image as `image/webp`.",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "roll_number"
            },
            "required": true,
            "description": "roll_number",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/unknown"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "put-rolls-{id}-contact-sheet",
        "summary": "Upload a roll contact sheet",
        "description": "Uploads a raw webp payload (request body is the file bytes), stores it in R2, and updates the roll.",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "roll_number"
            },
            "required": true,
            "description": "roll_number",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactSheetUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Film": {},
      "CreateFilmBody": {},
...
      "unknown": {},
      "ContactSheetUploadResponse": {
        "type": "object",
        "properties": {
          "contact_sheet_url": {
            "type": "string"
          }
        }
      },
      "Roll": {},
      "CreateRollBody": {
        "type": "object",
        "properties": {
          "roll_number": {
            "type": "string",
            "nullable": false
          },
          "camera_id": {
            "type": "string",
            "nullable": true
          },
          "film_id": {
            "type": "string",
            "nullable": true
          },
          "shot_at": {
            "type": "string",
            "nullable": true
          },
          "fridge_at": {
            "type": "string",
            "nullable": true
          },
          "lab_at": {
            "type": "string",
            "nullable": true
          },
          "lab_name": {
            "type": "string",
            "nullable": true
          },
          "scanned_at": {
            "type": "string",
            "nullable": true
          },
          "processed_at": {
            "type": "string",
            "nullable": true
          },
          "uploaded_at": {
            "type": "string",
            "nullable": true
          },
          "archived_at": {
            "type": "string",
            "nullable": true
          },
          "album_name": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "push_pull": {
            "type": "number",
            "nullable": true
          }
        }
      }
    },
...
  }
}